nit: update qwen25 and move the resolve_kv_seq to modeling utils
