+++ QEfficient/transformers/models/grok_1/modeling_grok1.py
@@ -17,8 +17,9 @@ from transformers.modeling_outputs import (
 from transformers.models.llama.modeling_llama import repeat_kv
 
 from QEfficient.customop.rms_norm import CustomRMSNormFunc
-from QEfficient.transformers.cache_utils import QEffDynamicCache, resolve_kv_seq_len
+from QEfficient.transformers.cache_utils import QEffDynamicCache
 from QEfficient.transformers.modeling_attn_mask_utils import _create_causal_mask
+from QEfficient.transformers.modeling_utils import resolve_kv_seq_len
 from QEfficient.transformers.models.llama.modeling_llama import qeff_apply_rotary_pos_emb
 from QEfficient.utils.constants import MIN_MASKED_ATTENTION_VALUE
