+++ QEfficient/utils/_utils.py
@@ -67,6 +67,40 @@ class DownloadRetryLimitExceeded(Exception):
     """
 
 
+def resolve_kv_seq_len(
+    past_key_value: Optional[Any],
+    layer_idx: int,
+    current_seq_len: int,
+    cache_position: Optional[torch.LongTensor] = None,
+) -> int:
+    """
+    Resolve KV sequence length for rotary embeddings with cache compatibility.
+
+    Use the current key sequence length as baseline, then grow it with:
+    - cache_position max (when provided)
+    - cache object reported length for the current layer
+    """
+    resolved_seq_len = current_seq_len
+    if cache_position is not None and isinstance(cache_position, torch.Tensor) and cache_position.numel() > 0:
+        resolved_seq_len = max(resolved_seq_len, int(cache_position.max().item()) + 1)
+
+    if past_key_value is None:
+        return resolved_seq_len
+
+    get_seq_length = getattr(past_key_value, "get_seq_length", None)
+    if get_seq_length is None:
+        return resolved_seq_len
+
+    try:
+        cache_seq_len = get_seq_length(layer_idx)
+    except TypeError:
+        cache_seq_len = get_seq_length()
+
+    if cache_seq_len is None:
+        return resolved_seq_len
+    return max(resolved_seq_len, int(cache_seq_len))
+
+
 def login_and_download_hf_lm(model_name, *args, **kwargs):
     logger.info(f"loading HuggingFace model for {model_name}")
     hf_token = kwargs.pop("hf_token", None)
